Refactor upcoming matches parsing with dedicated parser class
